backend/requirements.txt
fastapi==0.115.1
uvicorn==0.34.0
firebase-admin==6.8.0
SQLAlchemy==2.0.40
python-dotenv==1.1.0
psycopg2-binary==2.9.10
yfinance==0.2.58
requests==2.32.3
pandas==2.2.3
alembic==1.15.1

